Dra ut FiksClient i egen fil
